+++ src/serialization/deserializer.ts
@@ -1,4 +1,0 @@
-export interface Serializer<ModelT> {
-    deserialize(modelData: any): ModelT;
-    serialize(model: ModelT): any;
-}